Initial commit of emscripten support for PDCurses
This commit includes basic input and output functionality. I've run
all the tests in demos with the exception of the tui test, and they
largely work as expected. There are some differences between
emscripten and SDL with the timing at the beginning of ozdemo and the
shooting star in xmas that I need to look into. There is also an
error that comes up, `Identifier 'ExitStatus' has already been declared`
that comes up during tests, though it doesn't seem to keep things from
running. Wide character support, mouse support, clipboard support and
headless testing are the major tasks left to complete.