Update cursor to use transparency
The cursor overlays a cell and sets its textContent to match the
textContent of the cell. Previously, the cursor would blink by
alternating between setting its foreground and background color to match
the cell and inverting those colors. I've altered the blink-cursor
animation to set foreground and background color to transparent
during the phase where it was matching the cell's foreground and
background. The reason for the change is that my preliminary wide
character testing has shown that the cursor would obscure the portion of
any wide character that overlaps with an adjacent cell when the
cursor is at that adjacent cell's position. Using transparency allows
the full wide character to be visible.